requirements.txt
psycopg2-binary==2.9.7
Pillow==10.0.0
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.0.4
pyscard==2.0.7
boto3==1.29.7
python-dotenv==1.0.0